+++ Final/src/main/java/web/dto/BookListInfo.java
@@ -5,6 +5,7 @@ public class BookListInfo { // 판매부수정보
 	private int magazineNo; // 판매부수번호
 	private String month; // 보유 호수
 	private int circulation; // 보유 부수
+	private String sellerId; // 판매자아이디
 	
 	public int getMagazineNo() {
 		return magazineNo;
@@ -24,10 +25,17 @@ public class BookListInfo { // 판매부수정보
 	public void setCirculation(int circulation) {
 		this.circulation = circulation;
 	}
+	public String getSellerId() {
+		return sellerId;
+	}
+	public void setSellerId(String sellerId) {
+		this.sellerId = sellerId;
+	}
 	
 	@Override
 	public String toString() {
-		return "BookListInfo [magazineNo=" + magazineNo + ", month=" + month + ", circulation=" + circulation + "]";
+		return "BookListInfo [magazineNo=" + magazineNo + ", month=" + month + ", circulation=" + circulation
+				+ ", sellerId=" + sellerId + "]";
 	}
 	
 }
