+++ Final/src/main/java/web/dto/SellerInfo.java
@@ -6,10 +6,14 @@ public class SellerInfo { // 판매자 계정
 	private String sellerPw; // 비밀번호
 	private String sellerPhone; // 전화번호
 	private String sellerName; // 이름
-	private int locNo; // 판매위치 번호 (fk)
-	private String bigdomId; // 빅돔 아이디 (fk)
 	private String sellerImg; // 판매자 사진
 	private String sort; // 유저분류
+	
+	@Override
+	public String toString() {
+		return "SellerInfo [sellerId=" + sellerId + ", sellerPw=" + sellerPw + ", sellerPhone=" + sellerPhone
+				+ ", sellerName=" + sellerName + ", sellerImg=" + sellerImg + ", sort=" + sort + "]";
+	}
 	
 	public String getSellerId() {
 		return sellerId;
@@ -35,18 +39,6 @@ public class SellerInfo { // 판매자 계정
 	public void setSellerName(String sellerName) {
 		this.sellerName = sellerName;
 	}
-	public int getLocNo() {
-		return locNo;
-	}
-	public void setLocNo(int locNo) {
-		this.locNo = locNo;
-	}
-	public String getBigdomId() {
-		return bigdomId;
-	}
-	public void setBigdomId(String bigdomId) {
-		this.bigdomId = bigdomId;
-	}
 	public String getSellerImg() {
 		return sellerImg;
 	}
@@ -58,13 +50,6 @@ public class SellerInfo { // 판매자 계정
 	}
 	public void setSort(String sort) {
 		this.sort = sort;
-	}
-	
-	@Override
-	public String toString() {
-		return "SellerInfo [sellerId=" + sellerId + ", sellerPw=" + sellerPw + ", sellerPhone=" + sellerPhone
-				+ ", sellerName=" + sellerName + ", locNo=" + locNo + ", bigdomId=" + bigdomId + ", sellerImg="
-				+ sellerImg + ", sort=" + sort + "]";
 	}
 	
 }
