+++ Final/src/main/webapp/WEB-INF/tiles/layout_seller.jsp
@@ -21,12 +21,12 @@
 <style type="text/css">
 	/* 모든 페이지에 적용되는 스타일 */
 	
-	.wrap {
-		width: 1100px;
-		margin: 0 auto;
-		border: 1px solid #ccc;
-		padding-bottom: 20px;
-	}
+/* 	.wrap { */
+/* 		width: 1100px; */
+/* 		margin: 0 auto; */
+/* 		border: 1px solid #ccc; */
+/* 		padding-bottom: 20px; */
+/* 	} */
 </style>
 
 <script type="text/javascript">
@@ -38,10 +38,12 @@
 <body>
 
 <!-- HEADER -->
-<tiles:insertAttribute name="header" />
+<div class="container">
+	<tiles:insertAttribute name="header" />
+</div>
 
 <!-- BODY -->
-<div class="wrap">
+<div class="wrap container">
 	<tiles:insertAttribute name="body" />
 </div>
